+++ tests/unit/web/RequestTest.php
@@ -386,6 +386,23 @@ class RequestTest extends TestCase
         self::assertFalse($this->getInaccessibleProperty($this->request, '_isActionRequest'));
     }
 
+    public function testGetPreviewParam(): void
+    {
+        self::assertNull($this->request->getPreviewParam());
+
+        $this->request->setQueryParams(['x-craft-preview' => 'foobar']);
+        self::assertEquals('foobar', $this->request->getPreviewParam());
+
+        $this->request->setQueryParams([]);
+        self::assertNull($this->request->getPreviewParam());
+
+        $this->request->setQueryParams(['x-craft-live-preview' => 'foobar']);
+        self::assertEquals('foobar', $this->request->getPreviewParam());
+
+        $this->request->setQueryParams([]);
+        self::assertNull($this->request->getPreviewParam());
+    }
+
     /**
      * @dataProvider acceptsDataProvider
      */